+++ NEWS.md
@@ -1,6 +1,6 @@
 # gsignal 0.3-3
 
-- date: 20220328
+- date: 20220330
 - Fixed Github Issue #3: Problems with fftfilt when FFT length is provided by user
 - copy attributes of input object x to output in functions filter, filtfilt, sosfilt, fftfilt
 - copy dimnames of input object x to output in functions upfirdn, resample, upsample, upsamplefill,
@@ -14,7 +14,7 @@
 - adapted ifft() to use isConjSymm instead of ZapIm
 - reduced default tolerance for isWhole() and zapIm()
 - bugfix in detrend(): function now returns a vector if input was a vector
-- bugfix in filtfilt
+- bugfix in filtfilt(): corrected bug in computing filter ends (default and Sos methods)
 
 
 # gsignal 0.3-2